+++ src/cerebrus/base/__init__.py
@@ -1,3 +1,2 @@
 from cerebrus.base.cerebro import ICerebro
 from cerebrus.base.parser import Parser
-from cerebrus.base.qeeg import IPowerSpectralAnalysis