+++ src/cerebrus/base/cerebro.py
@@ -31,3 +31,16 @@ class ICerebro(abc.ABC):
         Analyze the preprocessed data.
         """
         pass
+
+    @abc.abstractmethod
+    def write_json(self, filename: str) -> None:
+        """
+        Save stored analysis data in JSON file format.
+        """
+        pass
+
+    @abc.abstractmethod
+    def write_hdf5(self, filename: str) -> None:
+        """
+        Save stored analysis jdata in hdf5 format.
+        """